Prepare PDFium for updated base::span implementation.
There are several blocking issues:
-- subspan(), first(), and last() insist upon strict size_t arguments
-- Fixed-size spans must be extracted when passing as arguments.
-- Some CTAD not working for arguments to reinterpret_span.
-- as_bytes() has a safety check which requires an alternative form.

To resolve these:
-- convert to templated subspan(), first(), and last().
-- add 'u' to constants where otherwise avoid the above.
-- static_cast<> when all else fails
-- add some explict span() calls.
-- implement stub for two-arg form of as_bytes()

There will likely be some small follow-ups required to catch platform
specific issues.

Bug: 42271100
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/131050
Commit-Queue: Tom Sepez <tsepez@chromium.org>

+++ core/fpdfapi/font/cpdf_cmap.cpp
@@ -356,9 +356,9 @@ uint32_t CPDF_CMap::GetNextChar(ByteStringView pString, size_t* pOffset) const {
       int char_size = 1;
       codes[0] = offset < pBytes.size() ? pBytes[offset++] : 0;
       while (true) {
-        int ret =
-            CheckFourByteCodeRange(pdfium::make_span(codes).first(char_size),
-                                   mixed_four_byte_leading_ranges_);
+        int ret = CheckFourByteCodeRange(
+            pdfium::make_span(codes).first(static_cast<size_t>(char_size)),
+            mixed_four_byte_leading_ranges_);
         if (ret == 0) {
           return 0;
         }
